requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.3
fastapi-cli==0.0.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.51.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
openai==1.52.0
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rpds-py==0.20.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.0
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.5
twilio==9.3.6
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
yarl==1.16.0
zipp==3.20.2
anthropic==0.39.0
opentelemetry-sdk < 1.30.0
logfire[fastapi]==3.5.1
typeguard==4.4.1
pytest==8.3.4